+++ CHANGES.rst
@@ -1,6 +1,14 @@
 0.6 (unreleased)
 ----------------
 
+General
+^^^^^^^
+
+- ``photutils.psf``
+
+  - Added ``single_object_model`` options to BasicPSFPhotometry and 
+    IterativelySubtractedPSFPhotometry.
+
 API changes
 ^^^^^^^^^^^
 
@@ -12,6 +20,12 @@ API changes
 
   - The ``find_peaks`` function will no longer issue a RuntimeWarning
     if the input data contains NaNs. [#712]
+
+- ``photutils.funcs``
+
+  - Added ``SingleObjectModelBase`` abstract class, and ``SingleObjectModel`` 
+    class, which maintain backwards compatibility with previous versions in
+    ``photutils.psf.BasicPSFPhotometry`` and subclasses.
 
 - ``photutils.epsf``